fix(core): Updates VecEncoder to support anything that can be bytes
This came from an issue when trying to pass a vec of tuples that had
`Bytes` ended up as `&&Bytes` which there wasn't an impl for. When I
looked at the implementation, it looked like pretty much everything was
calling `AsRef` anyway, so I collapsed everything down to the single
impl.

Signed-off-by: Taylor Thomas <taylor@cosmonic.com>